safari-600.7-branch/LayoutTests/fast/mediastream/MediaStream-add-ended-tracks.html

<!DOCTYPE html>
<html>
    <head>
        <script src="../../resources/js-test-pre.js"></script>
        <script>

            var stream1;
            var stream2;
            var audioTrack;

            function error()
            {
                testFailed('Stream generation failed.');
                finishJSTest();
            }

            function getUserMedia(dictionary, callback)
            {
                try {
                    navigator.webkitGetUserMedia(dictionary, callback, error);
                } catch (e) {
                    testFailed('webkitGetUserMedia threw exception :' + e);
                    finishJSTest();
                }
            }

            function tryAddTrack(stream, track)
            {
                try {
                    stream.addTrack(track);
                } catch (exception) {
                    testFailed("addTrack threw an exception.");
                    finishJSTest();
                }
            }

            function createStreamAndAddTracks()
            {
                shouldBe('stream2.getAudioTracks().length', '1');
                audioTrack = stream1.getAudioTracks()[0];
                audioTrack.stop();
                shouldBeEqualToString('audioTrack.readyState', 'ended');
                tryAddTrack(stream2, audioTrack);
                shouldBe('stream2.getAudioTracks().length', '2');

                finishJSTest();
            }

            function gotStream2(s)
            {
                stream2 = s;
                createStreamAndAddTracks();
            }

            function gotStream1(s)
            {
                stream1 = s;
                shouldBe('stream1.getAudioTracks().length', '1');
                getUserMedia({audio:true, video:true}, gotStream2);
            }

            function startMedia()
            {
                description("Test adding tracks to inactive MediaStream.");
                getUserMedia({audio:true, video:true}, gotStream1); 
            }
            window.jsTestIsAsync = true;
            window.successfullyParsed = true;
        </script>
        <script src="../../resources/js-test-post.js"></script>
    </head>
    <body onload="startMedia()">
        <p id="description"></p>
        <div id="console"></div>
    </body>
</html>
